move the creation of digital subscription and supporterplus test products into a shared directory
