+++ modules/zuora/test/fixtures/request-bodies/digitalSub-subscribe-body-old-price.ts
@@ -1,6 +1,6 @@
 import { zuoraDateFormat } from '@modules/zuora/common';
 import type { Dayjs } from 'dayjs';
-import { catalog } from '../../../src/productToDiscountMapping';
+import { catalog } from '../../../../../handlers/discount-api/src/productToDiscountMapping';
 
 export const digiSubSubscribeBody = (
 	subscriptionDate: Dayjs,
